move d2d bitmap cache into its own file
