+++ vstgui/lib/platform/win32/direct2d/d2dbitmap.h
@@ -72,26 +72,6 @@ protected:
 	IWICBitmapSource* source;
 };
 
-//-----------------------------------------------------------------------------
-class D2DBitmapCache
-{
-public:
-	ID2D1Bitmap* getBitmap (D2DBitmap* bitmap, ID2D1RenderTarget* renderTarget);
-	
-	void removeBitmap (D2DBitmap* bitmap);
-	void removeRenderTarget (ID2D1RenderTarget* renderTarget);
-
-	static D2DBitmapCache* instance ();
-//-----------------------------------------------------------------------------
-protected:
-	D2DBitmapCache ();
-	~D2DBitmapCache ();
-	ID2D1Bitmap* createBitmap (D2DBitmap* bitmap, ID2D1RenderTarget* renderTarget);
-	using RenderTargetBitmapMap = std::map<ID2D1RenderTarget*, ID2D1Bitmap*>;
-	using BitmapCache = std::map<D2DBitmap*, RenderTargetBitmapMap>;
-	BitmapCache cache;
-};
-
 } // VSTGUI
 
 #endif // WINDOWS
