+++ vstgui/vstgui_win32.cpp
@@ -23,6 +23,7 @@
 #include "lib/platform/win32/winstring.cpp"
 #include "lib/platform/win32/wintimer.cpp"
 #include "lib/platform/win32/direct2d/d2dbitmap.cpp"
+#include "lib/platform/win32/direct2d/d2dbitmapcache.cpp"
 #include "lib/platform/win32/direct2d/d2ddrawcontext.cpp"
 #include "lib/platform/win32/direct2d/d2dfont.cpp"
 #include "lib/platform/win32/direct2d/d2dgradient.cpp"
